libtirpc: Don't install FreeBSD and Solaris headers
Only install headers specific to tirpc.

There are some extra headers in the libtirpc tarball that are just
for the library to build.  They are required because libtirpc is
a port from FreeBSD (and before that, Solaris).

These headers are not meant to be installed in /usr/include, and do not
exist on the reference implementation (ie. Solaris).

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Steve Dickson <steved@redhat.com>