Add xdr_dplx_msg_decode_(start|continue).
These routines provide an interface to decode an incoming
CALL or REPLY in two parts.  The direction and xid (and for
calls requested call specifics) are known after the start
call, which permits a dispatch step for concurrent callers
and service handlers.  More complete inlining also happens
in the split calls.