Lock-free encoding and independent svc reply queue.
Implements an out-of-line, blocking reply queue for svc_vc
transports.

It -would- appear logical to add an xdr_ioq_truncate operation
permitting a caller to discard bytes at the end of the stream, if
the need for it arose.

Works, maybe consolidate vrec->off and vrec->len, which now appear the
same.

Signed-off-by: Matt Benjamin <matt@linuxbox.com>

Conflicts:
	src/clnt_vc.c
	src/svc_vc.c

Signed-off-by: Matt Benjamin <matt@linuxbox.com>