Use configured sendsize + 8K for ioq max segment size.
Pending support for GSSAPI iov interfaces, svc_vc_reply ecodes
in one buffer segment.  Since most replies are small, an initial
buffer segment of 8K ("page size") is tried first.  If a reply
would exceed this, it is reallocated with a larger size, which
with this change, is of length MaxRPCSendBufferSize + 8K.

Signed-off-by: Matt Benjamin <matt@linuxbox.com>