Add XDR x_flags and XDR_FLAG_CKSUM.
Add a shared flags member to XDR, and use it for fine-grained
control over request checksum behavior.

Internal checksumming when decoding is an acceptable default
since it reduces work for both NFSv3 and NFSv4.0.  Since the
checksum is unecessary for NFSv4.1, allow the caller to omit
the checksum step by clearing the approprite bit in XDR
x_flags.

Signed-off-by: Matt Benjamin <matt@linuxbox.com>