Add a pthread key initializer constant
Clean up: replace the naked "-1" with a symbolic constant that helps
document what is going on.  The name matches the name of the other
pthread initializer constants.

Also, since pthread_key_t is an unsigned integer, use a type cast to
eliminate the implicit cast that occurs every time foo_key is
compared to -1.  This eliminates a number of compiler warnings.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Steve Dickson <steved@redhat.com>