man: Replace outdated Solaris and BSD references in the man pages
The .Lb macro in these man pages lists "libc" as the containing
library, which is not correct for Linux.

However, our .Lb macro does not appear to support the value
"libtirpc."  List the containing library in an AVAILABILITY section instead.

Lastly, remove references to man pages that do not exist on Linux.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Steve Dickson <steved@redhat.com>