Rewrite inline reply/call
Implement inline encoding and decoding of RPC replies, since those were
unimplemented before.

In doing so, I merged the inline and out-of-line encoding steps for both
call and reply, since we know how big everything is in advance.

I also had to change the amount grabbed for decoding, specifically we
grab five words (which is the amount we're guaranteed for replies) and
then three more later (to fill out the header for calls.)

This fixes one part of the layoutrecall bug reported by Boaz.

Signed-off-by: Adam C. Emerson <aemerson@linuxbox.com>