Remove now-obsolete auth_get() in authunix_ncreate().
This change is intended to be equivalent a proposed change by
Kelong Wang <buaawkl@gmail.com>, which restored the default
value of ah_refcnt to 0.  Speculatively, the change this reverted
was incomplete, but an impulse may have been to avoid the
atomic increment of ah_refcnt, which is known only to the
caller.

Signed-off-by: Matt Benjamin <matt@cohortfs.com>