Refactor authgss_refresh()
rpc_gss_seccreate() can report a rich set of information about
the newly created GSS context, if the caller provides the
"options_ret" argument.

This argument has to be available to authgss_refresh() so that
it can be filled in properly.

This appears to be the "official" way of doing what
authgss_get_private_data() does now, though I suspect it's not a
perfect match.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Steve Dickson <steved@redhat.com>