rpc_broadcast: handle misformed rpcbind replies
Some rpcbind implementations seem to return IPv6 uaddrs
in response to an IPv4 broadcast (which is probably due
to their using a single v6 socket to handle both v6 and
v4 requests).

We can either discard these replies, or fix them up silently.
Here's a patch that implements the latter.

Signed-off-by: Olaf Kirch <okir@suse.de>
Signed-off-by: Steve Dickson <steved@redhat.com>