Symbol versioning patch
Added the functions, which are not marked
static and for which we provide a protoype
in the public installed header files.

Additional, there is the function __libc_clntudp_bufcreate
for glibc compatibility and the three functions:
__svc_clean_idle; svc_auth_none; libtirpc_set_debug
used by rpcbind.

I have added them as "TIRPC_PRIVATE". If we want to provide them
as official functions for everybody, we should add prototypes
to the header files and move them to the TIRPC_0.3.0 section.

Signed-off-by: Steve Dickson <steved@redhat.com>