blkin: introduce blkin/zipkin tracing
The enclosing CMake environment should have defined HAVE_BLKIN,
and provide for linkage with libblkin and liblttng-ust.

When HAVE_BLKIN is defined, a blkin endpoint is defined in the
SVCXPRT handle.  The endpoint is used whenever tracepoints are
emitted.

The Zipkin framework assumes that the all addresses are ipv4, and
can be stored as integers.  To get around this, ignore the address
field and format the address info into the endpoint name.

The endpoint name needs to persist for the lifetime of the xprt
handle, so provide for that.  (The blkin variables are all kept in
an anonymous 'blkin' struct in the handle.)

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>