xdr_inline invalid memory access:
 - inline_xdr_string tries to compute strlen on strings that are not
necessarily valid, for a value we do not need
 - mem_free does not use its second argument, which already was invalid
in inline_xdr_bytes and now is in inline_xdr_free. mem_free has a comment
cautioning about using second argument where it is defined, hopefully -1
should be easier to spot than some uninitialized variable if someone ever
tries to change it.

Signed-off-by: Dominique Martinet <dominique.martinet@cea.fr>