If available, use the GNUInstallDirs module to determine libdir name.
GNUInstallDirs takes care of a lot of tedious path setting for us but is
not available in older versions of CMake. Use it if available, otherwise
fall back to the old way of determining LIB_INSTALL_DIR.

This will be useful for more thing in the future, for now let's stick
with LIB_INSTALL_DIR to keep the commit minimal.