At the time of this writing, glibc-2.26 has a bug where   sysconf(_SC_IOV_MAX); always returns -1.
Versions known to have this bug are glibc-2.26-8 in Fedora 27 beta
and glibc-2.26.90-16 in Fedora 28 rawhide.

Upstream glibc BZ at
  https://sourceware.org/bugzilla/show_bug.cgi?id=22321

Linux does have a limit on the size of the iovec that is passed in
readv/writev syscalls; defined as UIO_MAXIOV (i.e. 1024) in
.../include/uapi/linux/uio.h. See kernel source readv/writev syscalls.

It's also possible to get _XOPEN_IOV_MAX (i.e. 16) and IOV_MAX (1024)
by wrapping <limits.h> with #define __USE_XOPEN ... #undef but that's
XOpen and theoretically Linux could change.