LTTng support
Support tracing with LTTng in ntirpc.  This gets complicated when ntirpc
is a submodule (as it is in Ganesha), hence the define USE_NTIRPC_LTTNG
being different than the option USE_LTTNG.

To use this in a program, link against ntirpc normally, then link in the
static library libntirpc_lttng.a to get the symbols defined.  You will
need to PRELOAD libntirpc_tracing.so when you want tracing to actually
happen, just like normal.

Signed-off-by: Daniel Gryniewicz <dang@redhat.com>