Add ability to suspend requests
If the protocol layer needs to save some data (it probably does), it
should use SVCXPRT->xp_u1 or SVCXPRT->xp_u2 (Ganesha only uses xp_u2
so it could use xp_u1).

If a request is to be suspended, the SVCXPRT and svc_req should not be
used once it is possible for another thread to be processing an async
completion of the request. This is because if such an async completion
happened to run before the original thread returned up the stack, the
request could actually be requeued and even executing on another RPC
thread while the first thread is still unwinding.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>