Force urcu-bp for now
When _LGPL_SOURCE is defined the lttng headers will inline some of the
RCU functions with the "bulletproof" functions. This means that we can't
currently select a urcu flavor at runtime that will be used universally.
I'm a little unclear on what happens if you mix different flavored urcu
calls, but I doubt that it's anything good.

Let's switch back to using urcu-bp universally for now until this has a
better solution. It's not ideal for performance, but our use of URCU is
pretty minimal for now, and it's better to be safe than fast.

Signed-off-by: Jeff Layton <jlayton@redhat.com>