Fix XPRT leak for TCP connections
XPRT refcounting could leak a ref in some cases.  Fix this by dropping
the extra ref in error cases, and by passing the event ref (which should
have been dropped, because events are oneshot) to the event processor.

This fixes the FD depletion while doing mount/unmount in a tight loop.

Signed-off-by: Daniel Gryniewicz <dang@redhat.com>