Make UDP XPRT refs match TCP
TCP has a double-ref setup, where the tree has one, and the event loop
has one.  This allows processors to destroy the xprt on error without
freeing it, ensuring no use-after-free.

UDP does not have a XPRT tree, since it re-uses a single socket, so it
only has one ref.  This means that destroy-on-error immediately frees.

Add an extra ref to the UDP path so that it matches TCP, and to allow
destroy.

Signed-off-by: Daniel Gryniewicz <dang@redhat.com>