xdr_ioq.c: fix xdr_ioq_setpos to deal with partially full buffers
If a vector includes some partially filled buffers, setpos must
position into the next buffer if it would land at or past the tail
of a buffer that is not the terminal buffer.

This will be required if putbufs is made to work.

It is also required for gss_wrap_iov where we need to insert a header
just before the response data after the response data has already been
encoded. This means the response data MUST start in a new buffer, leaving
the previous buffer potentially partially full.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>