_svcauth_gss - Clean up gss_data handling
gss_data (gd) is refcounted.  authgss_ctx_hash_get() takes a ref,
alloc_svc_rpc_gss_data() takes a ref, and authgss_ctx_hash_set() takes a
ref.  This means that gd needs to be unref'd on failure, always.

In addition, remove gd_hashed, since we need to unref if we allocated
it, and it cannot be hashed for RPCSEC_GSS_INIT or
RPCSEC_GSS_CONTINUE_INIT.

Signed-off-by: Daniel Gryniewicz <dang@fprintf.net>