Implement nonblocking send - convert from writev to sendmsg
The xdr_ioq keeps track of the progress of send and if any calls to
sendmsg ever would have blocked. sendmsg will send as much data as it
can without blocking, and then a subsequent call to send more data will
likely return EAGAIN or EWOULDBLOCK to indicate it might block. A
pointer to the duplex record is also kept for ease of finding it when
the epoll event to unblock is triggered.

Each duplex record now has two events, one for receive and one for
send. It also has a pointer to the in progress send xdr_ioq.

The SVCXPRT has a duplicate xp_fd_send to be used along with event_send
to allow epoll for both receive and waiting for send to unblock.

svc_ioq_flushv has the most changes to manage the conversion to sendmsg
and dealing with partial write of the xdr_ioq.

With the possibility of blocked send xdr_ioq, the writeq must be
managed differently. Every xdr_ioq is queued into the writeq, but
before doing so we remember if the writeq was empty so we can either
immediately start working on the send (svc_ioq_write_now) or submit a
work pool task (svc_ioq_write_submit). When the send unblocks,
svc_rqst_xprt_task_send directly calls svc_ioq_write since the blocked
send is at the head of the writeq.

Signed-off-by: Daniel Gryniewicz <dang@redhat.com>
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>