Fix xprt refcounting for non-blocking write
Nonblocking write was missing a ref in the case where the first entry in
the queue finished, causing the epoll to be unhooked.  If the next entry
blocked, then epoll is hooked again, but the hook path didn't take a
ref, causing an extra unref later.  This destroys the xprt, closing the
connection, and hanging the client.

Signed-off-by: Daniel Gryniewicz <dang@redhat.com>