Define a presumed UIO_MAXIOV value as needed
Do not use UIO_MAXIOV directly except in one place, in order to
accommodate macOS, which only defines it in the kernel version of
sys/uio.h.

Define an adapter value which is portable between platforms as long as
svc_internal.h is included.

An alternative solution would be to define UIO_MAXIOV ourselves if it is
missing.  I chose against this since it is somewhat obfuscated, but I am
not strongly attached to this style.

Signed-off-by: Matthew DeVore <matvore@google.com>