Centrally define bswap_## on non-GNU platforms
There is some redundancy in Ganesha's codebase around defining
portability shims for this functionality (e.g. [0] and [1]), so it would
be nice to store them in one location. This also allows more portions of
Ganesha to build on macOS, though the files will have to be updated to
include <misc/portable.h> in order to build.

[0] src/support/city.c
[1] src/FSAL/FSAL_MEM/mem_export.c

Signed-off-by: Matthew DeVore <matvore@google.com>