Change xdrproc_t to only 2 parameters
The third parameter is always called with 0.

All the called functions, which have been cast to xdrproc_t only
take two parameters. Passing the extra 0 doesn't hurt anything but
could be a point of confusion and is extraneous code.

Also, add 2 dummy parameters to xdr_void which is the single function
cast to xdrproc_t that doesn't take two parameters.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>